Exctract first paint color scheme logic
The idea behind this improvement is to use user preferred color scheme in the
window between first paint and react initial render. This includes:

1. Running inlined javascripit that checks for preferred-color-scheme and
applies dark theme styles if preferred. It is critical to run this script
before react entrypoint script fetch to avoid flashing an unstyled document.
2. Clearing applied dark theme styles to allow toggling between light and dark
mode with mui theme styles.